Remove Chicken Leaper from planet 21, Vis Ragstrom
Fear the Leaper, man! Its weapon is misconfigured and deals silly impulse that instagibs striders.

In theory this would be grounds to fix Leaper (or to keep it just for Brutal) but also it doesn't actually leap, it has a bad animation, and there are already other melee range chickens. There's a reason it's not used for the main chicken defense mode.

Fixes ZeroK-RTS/Zero-K#4307